add icons in emojies place remove back button in fav,prof and hist pages

+++ app/src/main/java/com/example/moviesphere/DatabaseHelper.java
@@ -9,7 +9,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class DatabaseHelper extends SQLiteOpenHelper {
 
     private static final String DATABASE_NAME = "MovieSphere.db";
-    private static final int DATABASE_VERSION = 2; // Incremented version
+    private static final int DATABASE_VERSION = 5; 
 
     // Users Table
     private static final String TABLE_USERS = "users";
@@ -25,8 +25,8 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     private static final String COL_HISTORY_MOVIE_TITLE = "movie_title";
     private static final String COL_HISTORY_MOVIE_ID = "movie_id";
     private static final String COL_HISTORY_POSTER = "poster_url";
-    private static final String COL_HISTORY_YEAR = "year"; // New column
-    private static final String COL_HISTORY_TYPE = "type"; // New column
+    private static final String COL_HISTORY_YEAR = "year";
+    private static final String COL_HISTORY_TYPE = "type";
     private static final String COL_HISTORY_DATE = "date_added";
 
     // Favourites Table
@@ -38,6 +38,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     private static final String COL_FAV_POSTER = "poster_url";
     private static final String COL_FAV_RATING = "rating";
     private static final String COL_FAV_YEAR = "year";
+    private static final String COL_FAV_DATE = "date_added";
 
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -46,15 +47,14 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         // Create Users Table
-        String createUsersTable = "CREATE TABLE " + TABLE_USERS + " (" +
+        db.execSQL("CREATE TABLE " + TABLE_USERS + " (" +
                 COL_USER_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 COL_USERNAME + " TEXT UNIQUE NOT NULL, " +
                 COL_EMAIL + " TEXT UNIQUE NOT NULL, " +
-                COL_PASSWORD + " TEXT NOT NULL)";
-        db.execSQL(createUsersTable);
+                COL_PASSWORD + " TEXT NOT NULL)");
 
         // Create History Table
-        String createHistoryTable = "CREATE TABLE " + TABLE_HISTORY + " (" +
+        db.execSQL("CREATE TABLE " + TABLE_HISTORY + " (" +
                 COL_HISTORY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 COL_HISTORY_USER_ID + " INTEGER NOT NULL, " +
                 COL_HISTORY_MOVIE_TITLE + " TEXT NOT NULL, " +
@@ -62,12 +62,11 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 COL_HISTORY_POSTER + " TEXT, " +
                 COL_HISTORY_YEAR + " TEXT, " +
                 COL_HISTORY_TYPE + " TEXT, " +
-                COL_HISTORY_DATE + " DATETIME DEFAULT CURRENT_TIMESTAMP, " +
-                "FOREIGN KEY(" + COL_HISTORY_USER_ID + ") REFERENCES " + TABLE_USERS + "(" + COL_USER_ID + "))";
-        db.execSQL(createHistoryTable);
+                COL_HISTORY_DATE + " INTEGER, " +
+                "FOREIGN KEY(" + COL_HISTORY_USER_ID + ") REFERENCES " + TABLE_USERS + "(" + COL_USER_ID + "))");
 
         // Create Favourites Table
-        String createFavouritesTable = "CREATE TABLE " + TABLE_FAVOURITES + " (" +
+        db.execSQL("CREATE TABLE " + TABLE_FAVOURITES + " (" +
                 COL_FAV_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 COL_FAV_USER_ID + " INTEGER NOT NULL, " +
                 COL_FAV_MOVIE_TITLE + " TEXT NOT NULL, " +
@@ -75,18 +74,17 @@ public class DatabaseHelper extends SQLiteOpenHelper {
                 COL_FAV_POSTER + " TEXT, " +
                 COL_FAV_RATING + " TEXT, " +
                 COL_FAV_YEAR + " TEXT, " +
+                COL_FAV_DATE + " INTEGER, " +
                 "UNIQUE(" + COL_FAV_USER_ID + ", " + COL_FAV_MOVIE_TITLE + "), " +
-                "FOREIGN KEY(" + COL_FAV_USER_ID + ") REFERENCES " + TABLE_USERS + "(" + COL_USER_ID + "))";
-        db.execSQL(createFavouritesTable);
+                "FOREIGN KEY(" + COL_FAV_USER_ID + ") REFERENCES " + TABLE_USERS + "(" + COL_USER_ID + "))");
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        if (oldVersion < 2) {
-            // Add columns if upgrading to version 2
-            db.execSQL("ALTER TABLE " + TABLE_HISTORY + " ADD COLUMN " + COL_HISTORY_YEAR + " TEXT");
-            db.execSQL("ALTER TABLE " + TABLE_HISTORY + " ADD COLUMN " + COL_HISTORY_TYPE + " TEXT");
-        }
+        db.execSQL("DROP TABLE IF EXISTS " + TABLE_HISTORY);
+        db.execSQL("DROP TABLE IF EXISTS " + TABLE_FAVOURITES);
+        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
+        onCreate(db);
     }
 
     // ==================== USER METHODS ====================
@@ -97,7 +95,6 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         values.put(COL_USERNAME, username);
         values.put(COL_EMAIL, email);
         values.put(COL_PASSWORD, password);
-
         long result = db.insert(TABLE_USERS, null, values);
         db.close();
         return result != -1;
@@ -105,9 +102,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public boolean checkUser(String username, String password) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_USERS + " WHERE " +
-                COL_USERNAME + "=? AND " + COL_PASSWORD + "=?";
-        Cursor cursor = db.rawQuery(query, new String[]{username, password});
+        Cursor cursor = db.rawQuery("SELECT * FROM " + TABLE_USERS + " WHERE " + COL_USERNAME + "=? AND " + COL_PASSWORD + "=?", new String[]{username, password});
         boolean exists = cursor.getCount() > 0;
         cursor.close();
         db.close();
@@ -116,8 +111,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public boolean isUsernameExists(String username) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_USERS + " WHERE " + COL_USERNAME + "=?";
-        Cursor cursor = db.rawQuery(query, new String[]{username});
+        Cursor cursor = db.rawQuery("SELECT * FROM " + TABLE_USERS + " WHERE " + COL_USERNAME + "=?", new String[]{username});
         boolean exists = cursor.getCount() > 0;
         cursor.close();
         db.close();
@@ -126,8 +120,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public boolean isEmailExists(String email) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_USERS + " WHERE " + COL_EMAIL + "=?";
-        Cursor cursor = db.rawQuery(query, new String[]{email});
+        Cursor cursor = db.rawQuery("SELECT * FROM " + TABLE_USERS + " WHERE " + COL_EMAIL + "=?", new String[]{email});
         boolean exists = cursor.getCount() > 0;
         cursor.close();
         db.close();
@@ -136,8 +129,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public int getUserId(String username) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT " + COL_USER_ID + " FROM " + TABLE_USERS + " WHERE " + COL_USERNAME + "=?";
-        Cursor cursor = db.rawQuery(query, new String[]{username});
+        Cursor cursor = db.rawQuery("SELECT " + COL_USER_ID + " FROM " + TABLE_USERS + " WHERE " + COL_USERNAME + "=?", new String[]{username});
         int userId = -1;
         if (cursor.moveToFirst()) {
             userId = cursor.getInt(0);
@@ -151,75 +143,72 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public boolean addToHistory(int userId, String movieTitle, String movieId, String posterUrl, String year, String type) {
         SQLiteDatabase db = this.getWritableDatabase();
-
-        // Check if already exists
-        String checkQuery = "SELECT * FROM " + TABLE_HISTORY + " WHERE " +
-                COL_HISTORY_USER_ID + "=? AND " + COL_HISTORY_MOVIE_TITLE + "=?";
-        Cursor cursor = db.rawQuery(checkQuery, new String[]{String.valueOf(userId), movieTitle});
-
-        if (cursor.getCount() > 0) {
-            // Update the date if it already exists
-            ContentValues values = new ContentValues();
-            values.put(COL_HISTORY_DATE, "datetime('now')");
-            db.update(TABLE_HISTORY, values, COL_HISTORY_USER_ID + "=? AND " + COL_HISTORY_MOVIE_TITLE + "=?", 
+        db.beginTransaction();
+        try {
+            db.delete(TABLE_HISTORY, COL_HISTORY_USER_ID + "=? AND " + COL_HISTORY_MOVIE_TITLE + "=?", 
                     new String[]{String.valueOf(userId), movieTitle});
-            cursor.close();
-            db.close();
+
+            ContentValues values = new ContentValues();
+            values.put(COL_HISTORY_USER_ID, userId);
+            values.put(COL_HISTORY_MOVIE_TITLE, movieTitle);
+            values.put(COL_HISTORY_MOVIE_ID, movieId);
+            values.put(COL_HISTORY_POSTER, posterUrl);
+            values.put(COL_HISTORY_YEAR, year);
+            values.put(COL_HISTORY_TYPE, type);
+            values.put(COL_HISTORY_DATE, System.currentTimeMillis());
+
+            db.insert(TABLE_HISTORY, null, values);
+            db.setTransactionSuccessful();
             return true;
+        } finally {
+            db.endTransaction();
+            db.close();
         }
-        cursor.close();
-
-        ContentValues values = new ContentValues();
-        values.put(COL_HISTORY_USER_ID, userId);
-        values.put(COL_HISTORY_MOVIE_TITLE, movieTitle);
-        values.put(COL_HISTORY_MOVIE_ID, movieId);
-        values.put(COL_HISTORY_POSTER, posterUrl);
-        values.put(COL_HISTORY_YEAR, year);
-        values.put(COL_HISTORY_TYPE, type);
-
-        long result = db.insert(TABLE_HISTORY, null, values);
-        db.close();
-        return result != -1;
     }
 
     public Cursor getUserHistory(int userId) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_HISTORY + " WHERE " +
-                COL_HISTORY_USER_ID + "=? ORDER BY " + COL_HISTORY_DATE + " DESC";
-        return db.rawQuery(query, new String[]{String.valueOf(userId)});
+        return db.rawQuery("SELECT * FROM " + TABLE_HISTORY + " WHERE " + COL_HISTORY_USER_ID + "=? ORDER BY " + COL_HISTORY_DATE + " DESC", 
+                new String[]{String.valueOf(userId)});
     }
 
     // ==================== FAVOURITES METHODS ====================
 
     public boolean addToFavourites(int userId, String movieTitle, String movieId, String posterUrl, String rating, String year) {
         SQLiteDatabase db = this.getWritableDatabase();
-        ContentValues values = new ContentValues();
-        values.put(COL_FAV_USER_ID, userId);
-        values.put(COL_FAV_MOVIE_TITLE, movieTitle);
-        values.put(COL_FAV_MOVIE_ID, movieId);
-        values.put(COL_FAV_POSTER, posterUrl);
-        values.put(COL_FAV_RATING, rating);
-        values.put(COL_FAV_YEAR, year);
+        db.beginTransaction();
+        try {
+            db.delete(TABLE_FAVOURITES, COL_FAV_USER_ID + "=? AND " + COL_FAV_MOVIE_TITLE + "=?", 
+                    new String[]{String.valueOf(userId), movieTitle});
 
-        long result = db.insert(TABLE_FAVOURITES, null, values);
-        db.close();
-        return result != -1;
+            ContentValues values = new ContentValues();
+            values.put(COL_FAV_USER_ID, userId);
+            values.put(COL_FAV_MOVIE_TITLE, movieTitle);
+            values.put(COL_FAV_MOVIE_ID, movieId);
+            values.put(COL_FAV_POSTER, posterUrl);
+            values.put(COL_FAV_RATING, rating);
+            values.put(COL_FAV_YEAR, year);
+            values.put(COL_FAV_DATE, System.currentTimeMillis());
+
+            db.insert(TABLE_FAVOURITES, null, values);
+            db.setTransactionSuccessful();
+            return true;
+        } finally {
+            db.endTransaction();
+            db.close();
+        }
     }
 
     public boolean removeFromFavourites(int userId, String movieTitle) {
         SQLiteDatabase db = this.getWritableDatabase();
-        int result = db.delete(TABLE_FAVOURITES,
-                COL_FAV_USER_ID + "=? AND " + COL_FAV_MOVIE_TITLE + "=?",
-                new String[]{String.valueOf(userId), movieTitle});
+        int result = db.delete(TABLE_FAVOURITES, COL_FAV_USER_ID + "=? AND " + COL_FAV_MOVIE_TITLE + "=?", new String[]{String.valueOf(userId), movieTitle});
         db.close();
         return result > 0;
     }
 
     public boolean isFavourite(int userId, String movieTitle) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_FAVOURITES + " WHERE " +
-                COL_FAV_USER_ID + "=? AND " + COL_FAV_MOVIE_TITLE + "=?";
-        Cursor cursor = db.rawQuery(query, new String[]{String.valueOf(userId), movieTitle});
+        Cursor cursor = db.rawQuery("SELECT * FROM " + TABLE_FAVOURITES + " WHERE " + COL_FAV_USER_ID + "=? AND " + COL_FAV_MOVIE_TITLE + "=?", new String[]{String.valueOf(userId), movieTitle});
         boolean exists = cursor.getCount() > 0;
         cursor.close();
         db.close();
@@ -228,7 +217,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
     public Cursor getUserFavourites(int userId) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String query = "SELECT * FROM " + TABLE_FAVOURITES + " WHERE " + COL_FAV_USER_ID + "=?";
-        return db.rawQuery(query, new String[]{String.valueOf(userId)});
+        return db.rawQuery("SELECT * FROM " + TABLE_FAVOURITES + " WHERE " + COL_FAV_USER_ID + "=? ORDER BY " + COL_FAV_DATE + " DESC", 
+                new String[]{String.valueOf(userId)});
     }
 }
